# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Iremeapp REST APIs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/course/{school_id}": {
            "get": {
                "tags": [
                    "Get courses list API"
                ],
                "summary": "Get courses list API",
                "description": "Get courses list API",
                "operationId": "getCourseBySchool",
                "parameters": [
                    {
                        "name": "school_id",
                        "in": "path",
                        "description": "School ID",
                        "required": true
                    },
                    {
                        "name": "class_id",
                        "in": "query",
                        "description": "Class ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Courses are successfully fetched",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/course": {
            "post": {
                "tags": [
                    "Create new course API"
                ],
                "summary": "Create new course API",
                "description": "Create new course API",
                "operationId": "createNewCourse",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "school_id",
                                    "course_name",
                                    "course_id"
                                ],
                                "properties": {
                                    "school_id": {
                                        "type": "text"
                                    },
                                    "course_name": {
                                        "type": "text"
                                    },
                                    "course_id": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "New course is successfully created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/course/update/{school_id}": {
            "post": {
                "tags": [
                    "Update course API"
                ],
                "summary": "Update course API",
                "description": "Update course API",
                "operationId": "updateCourse",
                "parameters": [
                    {
                        "name": "school_id",
                        "in": "path",
                        "description": "School ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "course_name",
                                    "course_id"
                                ],
                                "properties": {
                                    "course_name": {
                                        "type": "text"
                                    },
                                    "course_id": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course is successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/course/destroy/{school_id}": {
            "post": {
                "tags": [
                    "Delete course API"
                ],
                "summary": "Delete course API",
                "description": "Delete course API",
                "operationId": "delete_Course",
                "parameters": [
                    {
                        "name": "school_id",
                        "in": "path",
                        "description": "School ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "course_id"
                                ],
                                "properties": {
                                    "course_id": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course is successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000//oauth/authorize",
                        "tokenUrl": "http://localhost:8000//oauth/token",
                        "refreshUrl": "http://localhost:8000//token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}